client/wayland: Generate library to be imported by panel
Wayland input-method protocol has set_overlay_panel to set
the IBus lookup window near by the text input cursor and
the calling Wayland APIs and calling GUI APIs need to be same PID.

Now `configure --enable-wayland` by default to include
ibuswaylandim.h in C files generated by Vala.

BUG=https://github.com/ibus/ibus/issues/2408
