+++ client/wayland/IBusWaylandIM-1.0.metadata
@@ -1,0 +1,1 @@
+IBusWaylandIM cheader_filename="ibuswaylandim.h"
